feat: improve SEO & loading speed on user profile
